[Refactor] Share issue @mention task routing across Gitea, GitLab, and GitHub (#611)
* refactor: share issue @mention orchestration across SCM providers

Extract provider-agnostic issue-mention env resolution, prompts, owner
lookup, and task start/follow-up routing so Gitea, GitLab, and GitHub
handlers stay thin adapters instead of near-duplicate monologues.

* refactor: keep issue-mention helpers internal to satisfy knip

---------
