+++ apps/api/src/handlers/gitea/__tests__/handleComment.test.ts
@@ -86,6 +86,10 @@ vi.mock('../../tasks/sendMessageToTask', () => ({
   steerMessageToTask: mockSteerMessageToTask,
 }));
 
+vi.mock('../../tasks/helpers', () => ({
+  findLatestTaskRun: vi.fn(),
+}));
+
 import { RunStatus, TaskPayloadKind } from '@roomote/types';
 
 import { handleGiteaComment } from '../handleComment';
