+++ apps/api/src/handlers/github/handleGitHubIssueComment.ts
@@ -1,38 +1,10 @@
-import {
-  buildMentionRequestBlock,
-  buildUntrustedContentPolicy,
-  buildUntrustedExternalContentBlock,
-  enqueueTask,
-  escapeTaskContextText,
-  getTaskUrl,
-} from '@roomote/cloud-agents/server';
-import {
-  db,
-  environmentRepositoryMappings,
-  eq,
-  asc,
-  findReusableGitHubIssueTaskOwner,
-} from '@roomote/db/server';
+import { getTaskUrl } from '@roomote/cloud-agents/server';
 import { getInstallationOctokit } from '@roomote/github';
-import {
-  type RunStatus,
-  type TaskPayload,
-  PRODUCT_NAME,
-  TaskPayloadKind,
-  isActivelyRunningTask,
-  isExitedRunStatus,
-} from '@roomote/types';
+import { PRODUCT_NAME } from '@roomote/types';
 
 import type { WebhookResponse } from '../../types';
-import {
-  buildSourceControlAccountLinkRequiredMessage,
-  buildSourceControlEnvironmentRequiredMessage,
-} from '../source-control-account-linking';
-import { findLatestTaskRun } from '../tasks/helpers';
-import {
-  sendMessageToTask,
-  steerMessageToTask,
-} from '../tasks/sendMessageToTask';
+import { buildSourceControlAccountLinkRequiredMessage } from '../source-control-account-linking';
+import { orchestrateIssueMention } from '../shared/issue-mention-orchestration';
 import { getGitHubAutomationTargets } from './getGitHubAutomationTargets';
 import { isMention } from './isMention';
 import type {
@@ -41,9 +13,6 @@ import type {
   WebhookInstallation,
   WebhookUser,
 } from './types';
-
-const EXISTING_TASK_WAIT_TIMEOUT_MS = 15_000;
-const EXISTING_TASK_WAIT_POLL_MS = 500;
 
 type IssueMentionIssue = {
   number: number;
@@ -66,83 +35,6 @@ type IssueMentionPayload = {
    */
   mentionBody?: string;
 };
-
-function buildIssueMentionPrompt({
-  repositoryFullName,
-  issueNumber,
-  issueTitle,
-  issueBody,
-  issueAuthorLogin,
-  issueUrl,
-  commentBody,
-  commenterLogin,
-}: {
-  repositoryFullName: string;
-  issueNumber: number;
-  issueTitle: string;
-  issueBody?: string | null;
-  issueAuthorLogin?: string | null;
-  issueUrl: string;
-  commentBody: string;
-  commenterLogin: string;
-}): string {
-  const trimmedIssueBody = issueBody?.trim() ?? '';
-  const trimmedCommentBody = commentBody.trim();
-  // `issues.opened` mentions arrive with the issue body as the mention text,
-  // so skip the duplicate issue-body context block when the two match.
-  const issueBodySection =
-    trimmedIssueBody && trimmedIssueBody !== trimmedCommentBody
-      ? [
-          '',
-          `Issue body (context only, authored by ${
-            issueAuthorLogin ? `@${issueAuthorLogin}` : 'an unknown user'
-          }):`,
-          buildUntrustedExternalContentBlock({
-            source: 'github_issue_body',
-            text: trimmedIssueBody,
-          }),
-        ]
-      : [];
-
-  return [
-    `${commenterLogin} mentioned Roomote on GitHub issue #${issueNumber} (${escapeTaskContextText(issueTitle)}) in ${repositoryFullName}.`,
-    `Issue URL: ${issueUrl}`,
-    '',
-    'Mention comment (the request to act on):',
-    buildMentionRequestBlock(trimmedCommentBody),
-    ...issueBodySection,
-    '',
-    buildUntrustedContentPolicy(),
-  ].join('\n');
-}
-
-function buildIssueFollowUpMessage({
-  repositoryFullName,
-  issueNumber,
-  issueTitle,
-  issueUrl,
-  commentBody,
-  commenterLogin,
-}: {
-  repositoryFullName: string;
-  issueNumber: number;
-  issueTitle: string;
-  issueUrl: string;
-  commentBody: string;
-  commenterLogin: string;
-}): string {
-  return [
-    `${commenterLogin} mentioned Roomote again on GitHub issue #${issueNumber} (${escapeTaskContextText(issueTitle)}) in ${repositoryFullName}.`,
-    `Issue URL: ${issueUrl}`,
-    '',
-    'This is a follow-up on the existing Roomote task for this issue. Continue that work instead of starting a separate task.',
-    '',
-    'Mention comment (the request to act on):',
-    buildMentionRequestBlock(commentBody),
-    '',
-    buildUntrustedContentPolicy(),
-  ].join('\n');
-}
 
 async function postIssueComment({
   installationId,
@@ -227,156 +119,6 @@ function buildStartFailedComment(): string {
   return `I saw the mention, but I could not start a task for this issue right now. Please try again in a moment.`;
 }
 
-async function resolveMappedEnvironmentId(
-  repositoryId: string,
-): Promise<string | null> {
-  const mappings = await db
-    .select({
-      environmentId: environmentRepositoryMappings.environmentId,
-    })
-    .from(environmentRepositoryMappings)
-    .where(eq(environmentRepositoryMappings.repositoryId, repositoryId))
-    .orderBy(asc(environmentRepositoryMappings.environmentId));
-
-  if (mappings.length === 0) {
-    return null;
-  }
-
-  // When multiple environments map to the same repository, pick a stable
-  // ordered environment id and still pin the selected repository so the worker
-  // has a concrete checkout target.
-  return mappings[0]?.environmentId ?? null;
-}
-
-async function waitForTaskToAcceptMessages({
-  taskId,
-  timeoutMs = EXISTING_TASK_WAIT_TIMEOUT_MS,
-}: {
-  taskId: string;
-  timeoutMs?: number;
-}) {
-  const deadline = Date.now() + timeoutMs;
-
-  while (Date.now() < deadline) {
-    const latestRun = await findLatestTaskRun(taskId, {
-      id: true,
-      status: true,
-      taskPhase: true,
-      sandboxServerUrl: true,
-    });
-
-    if (!latestRun || isExitedRunStatus(latestRun.status)) {
-      return null;
-    }
-
-    if (latestRun.sandboxServerUrl) {
-      return latestRun;
-    }
-
-    await new Promise((resolve) =>
-      setTimeout(resolve, EXISTING_TASK_WAIT_POLL_MS),
-    );
-  }
-
-  return null;
-}
-
-function isRetryableSandboxBootError(result: {
-  success: boolean;
-  status?: number;
-  error?: string;
-}): boolean {
-  if (result.success || result.status !== 409) {
-    return false;
-  }
-
-  const error = result.error ?? '';
-
-  return (
-    error.includes('no active sandbox') ||
-    error.includes('sandbox is still booting')
-  );
-}
-
-async function deliverIssueFollowUpToExistingTask({
-  taskId,
-  userId,
-  message,
-  status,
-  taskPhase,
-  commenterDisplayName,
-}: {
-  taskId: string;
-  userId: string;
-  message: string;
-  status: RunStatus;
-  taskPhase: string | null;
-  commenterDisplayName?: string;
-}) {
-  const deliver = ({
-    status,
-    taskPhase,
-  }: {
-    status: RunStatus;
-    taskPhase: string | null;
-  }) =>
-    isActivelyRunningTask(status, taskPhase)
-      ? steerMessageToTask({
-          taskId,
-          userId,
-          message,
-          senderMode: 'github_pr_follow_up',
-          ...(commenterDisplayName
-            ? { workerQuoteUserName: commenterDisplayName }
-            : {}),
-        })
-      : sendMessageToTask({
-          taskId,
-          userId,
-          message,
-          senderMode: 'github_pr_follow_up',
-          ...(commenterDisplayName
-            ? { workerQuoteUserName: commenterDisplayName }
-            : {}),
-        });
-
-  let attempt = await deliver({ status, taskPhase });
-
-  if (!isRetryableSandboxBootError(attempt)) {
-    return attempt;
-  }
-
-  const deadline = Date.now() + EXISTING_TASK_WAIT_TIMEOUT_MS;
-
-  while (Date.now() < deadline && isRetryableSandboxBootError(attempt)) {
-    // Prefer waiting until a sandbox URL is recorded, then keep retrying while
-    // delivery still reports sandbox-boot 409s (URL present, RPC not ready).
-    const readyRun = await waitForTaskToAcceptMessages({
-      taskId,
-      timeoutMs: Math.max(0, deadline - Date.now()),
-    });
-
-    if (!readyRun) {
-      return attempt;
-    }
-
-    attempt = await deliver({
-      status: readyRun.status,
-      taskPhase: readyRun.taskPhase,
-    });
-
-    if (!isRetryableSandboxBootError(attempt)) {
-      return attempt;
-    }
-
-    await new Promise((resolve) =>
-      setTimeout(resolve, EXISTING_TASK_WAIT_POLL_MS),
-    );
-  }
-
-  return attempt;
-}
-
 /**
  * Handle @mentions on plain GitHub issues (not pull requests).
  * Starts a standard task against the issue's repository, or continues an
@@ -458,137 +200,39 @@ export async function handleGitHubIssueComment(
     return { status: 'ok', message: 'account_link_required' };
   }
 
-  const environmentId = await resolveMappedEnvironmentId(target.repo.id);
-
-  if (!environmentId) {
-    await postIssueComment({
-      ...replyTarget,
-      body: buildSourceControlEnvironmentRequiredMessage('github'),
-    });
-
-    return { status: 'ok', message: 'environment_required' };
-  }
-
   const issueUrl =
     issue.html_url ??
     `https://github.com/${repositoryFullName}/issues/${issueNumber}`;
   const issueTitle = issue.title ?? `Issue #${issueNumber}`;
   const issueBody = issue.body ?? null;
   const commenterUserId = target.properties.userId;
+  const issueAuthorLogin = issue.user?.login ?? null;
 
-  const existingIssueOwner = await findReusableGitHubIssueTaskOwner({
-    repoFullName: repositoryFullName,
-    issueNumber,
-  });
-
-  if (existingIssueOwner?.taskId) {
-    const followUpMessage = buildIssueFollowUpMessage({
-      repositoryFullName,
-      issueNumber,
-      issueTitle,
-      issueUrl,
-      commentBody,
-      commenterLogin: sender.login,
-    });
-
-    const delivery = await deliverIssueFollowUpToExistingTask({
-      taskId: existingIssueOwner.taskId,
-      userId: commenterUserId,
-      message: followUpMessage,
-      status: existingIssueOwner.status,
-      taskPhase: existingIssueOwner.taskPhase,
-      commenterDisplayName: target.properties.githubLogin ?? sender.login,
-    });
-
-    if (delivery.success) {
-      await postIssueComment({
-        ...replyTarget,
-        body: formatFollowUpReply(
-          tryBuildTaskLink({
-            taskId: existingIssueOwner.taskId,
-            campaign: 'github.issue.mention.active-owner',
-          }),
-        ),
-      });
-
-      return { status: 'ok', message: 'active_issue_owner_routed' };
-    }
-
-    console.warn(
-      `[handleGitHubIssueComment] failed to deliver issue mention to reusable task ${existingIssueOwner.taskId}: ${delivery.error}`,
-    );
-  }
-
-  const prompt = buildIssueMentionPrompt({
+  return orchestrateIssueMention({
+    provider: 'github',
+    logPrefix: '[handleGitHubIssueComment]',
+    repositoryId: target.repo.id,
     repositoryFullName,
     issueNumber,
     issueTitle,
     issueBody,
-    issueAuthorLogin: issue.user?.login ?? null,
     issueUrl,
     commentBody,
     commenterLogin: sender.login,
+    commenterUserId,
+    githubLogin: target.properties.githubLogin ?? undefined,
+    githubUserId: target.properties.githubUserId ?? undefined,
+    followUpCommenterDisplayName: target.properties.githubLogin ?? sender.login,
+    retrySandboxBoot: true,
+    providerDisplayName: 'GitHub',
+    issueBodySource: 'github_issue_body',
+    issueBodyContextLabel: `Issue body (context only, authored by ${
+      issueAuthorLogin ? `@${issueAuthorLogin}` : 'an unknown user'
+    }):`,
+    postComment: (body) => postIssueComment({ ...replyTarget, body }),
+    formatFollowUpReply,
+    formatStartedReply,
+    formatStartFailed: buildStartFailedComment,
+    tryBuildTaskLink,
   });
-
-  const payload = {
-    repo: repositoryFullName,
-    environmentId,
-    selectedRepositories: [repositoryFullName],
-    description: prompt,
-    linkedWorkItems: [
-      {
-        provider: 'github',
-        identifier: String(issueNumber),
-        url: issueUrl,
-        title: issueTitle,
-        repository: repositoryFullName,
-      },
-    ],
-  } satisfies TaskPayload<typeof TaskPayloadKind.StandardTask>;
-
-  try {
-    const launch = await enqueueTask({
-      task: {
-        type: TaskPayloadKind.StandardTask,
-        githubLogin: target.properties.githubLogin,
-        githubUserId: target.properties.githubUserId,
-        payload,
-      },
-      initiator: { kind: 'user', userId: commenterUserId },
-      workflow: 'standard',
-      surface: 'github',
-      trigger: 'message',
-    });
-
-    await postIssueComment({
-      ...replyTarget,
-      body: formatStartedReply(
-        tryBuildTaskLink({
-          taskId: launch.taskId,
-          campaign: 'github.issue.mention',
-        }),
-      ),
-    });
-
-    return {
-      status: 'ok',
-      metadata: { ids: [launch.id] },
-    };
-  } catch (error) {
-    console.warn(
-      `[handleGitHubIssueComment] failed to start issue task for ${repositoryFullName}#${issueNumber}: ${
-        error instanceof Error ? error.message : String(error)
-      }`,
-    );
-
-    await postIssueComment({
-      ...replyTarget,
-      body: buildStartFailedComment(),
-    });
-
-    return {
-      status: 'error',
-      message: `issue_task_start_failed:${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
 }
